Feat: key derivation for libp2p identities
Keeping track of identities when setting up gateways is a pain.

This allows choosing a single secret for all gateways in a deployment, and
then deriving individual keys based on a given index. It also allows painless
key rotation.

Once we add bootstrap/peering capabilities, this functionality will be
extended and gateways will generate peer IDs for other gateways, look them up
on the DHT and peer with them automatically. That is, we will use the DHT for
service discovery, skipping manual peering configuration altogether.