feat: purge peer connections and information (#194)
* feat: purge peer connections and information

Connections to a specific peer, or to all peers, can be closed and the peer information removed from the peer store. This can be useful to help determine if the presence/absence of a connection to a peer is affecting behavior. Be aware that purging a connection is inherently racey as it is possible for the peer to reestablish a connection at any time following a purge.

* Output peers list in json
* Do not export mgr handler functions
* Update README.md